initial commit (johor-map)
Generate a component for a responsive map with hover and onclick feature.

To generate a map with styles
Use <svg><polygon>. I tried using <map> <area> but is unable to change the style because it is not a visible element.

(Needed to change) => I must generate a angular/material component with Grid-list schematic, and paste the codes i've done.